requirements.txt
azure-search-documents==11.7.0b2
azure-identity
openai
python-dotenv